# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Seans-TypeScript-NodeJS-CRUD-REST-API-Boilerplate",
        "description": "A minimal and easy to follow example of what you need to create a CRUD style API in NodeJs using TypeScript",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Dev, or from Heroku"
        },
        {
            "url": "/api/",
            "description": "With docker-compose and nginx proxy"
        }
    ],
    "tags": [
        {
            "name": "Jalves",
            "description": "API for Jalves Espetaria in the system"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/sessions": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a token in system",
                "requestBody": {
                    "description": "Authentication Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/authenticationRequest"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authentication"
                        }
                    },
                    "400": {
                        "description": "Incorrect email/password combination."
                    }
                }
            }
        },
        "/user/create":{
            "post":{
                "tags": [
                    "User"
                ],
                "summary": "Create a user in system",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/userRequest"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "400": {
                        "description": "Email address already used."
                    }
                }
            }
        },
        "/user/update/{id}":{
            "put":{
                "summary": "Update User with ID",
                "tags": [
                    "User"
                ],
                "requestBody":{
                    "description": "User Update",
                    "required": true,
                    "content": {
                        "aplication/json": {
                            "schema": {
                                "$ref": "#/definitions/userRequestUpdate"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "User with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "400": {
                        "description": "Email address already used."
                    },
                    "401": {
                        "description": "You do not permission update this user"
                    }
                }
            }
        },
        "/user/search/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user with given ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Cat with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "400": {
                        "description": "User do not exist!"
                    }
                }
            }
        },
        "/user/users/search": {
            "get": {
                "tags": [
                    "User"
                ],
                "security": [{"bearerAuth":[]}],
                "summary": "Get users with name and page",
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "Cat with name",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "Cat with page",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "400": {
                        "description": "User do not exist!"
                    },
                    "401": {
                        "description": "JWT token is missing!"
                    }
                }
            }
        }
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "userRequest":{
            "type": "object",
            "properties": {
                "name": {
                    "type":"string"
                },
                "email":{
                    "type":"string"
                },
                "age": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user":{
            "type":"object",
            "properties": {
                "name": {
                    "type":"string"
                },
                "email":{
                    "type":"string"
                },
                "age": {
                    "type": "integer"
                },
                "user_id":{
                    "type": "string"
                },
                "created_at":{
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "authenticationRequest":{
            "type":"object",
            "properties": {
                "email":{
                    "type":"string"
                },
                "password":{
                    "type":"string"
                }
            }
        },
        "authentication":{
            "type":"object",
            "properties": {
                "user":{
                    "type":"object",
                    "additionalProperties": {
                        "$ref": "#/definitions/user"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "userRequestUpdate": {
            "properties": {
                "name": {
                    "type":"string"
                },
                "email":{
                    "type":"string"
                },
                "age": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
    }
}